Move loader to own python file and added cmd flag to suppress its output
